Move hard link helper functions to their own class
